requirements.txt
django-tastypie==0.11.1
Django==1.6.4